+++ NEWS.md
@@ -1,6 +1,7 @@
 # confcons 0.2.0
 
 * `thresholds()` can now calculate information-based thresholds besides the mean-based thresholds. Two new parameters (`type` and `range`) are added to the function.
+* `consistence()` is renamed to `consistency()`
 
 # confcons 0.1.0